requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
nibabel>=5.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0
openpyxl>=3.1.0
